fastapi/requirements.txt
fastapi
uvicorn
google-cloud-storage
python-dotenv
requests
